fix: fix some files
