+++ js-plugins/vue/package.json
@@ -10,8 +10,7 @@
   },
   "exports":{
     ".":{
-      "import":"./index.mjs",
-      "require":"./index.cjs"
+      "import":"./esm/farm-vue-plugin.js"
     }
   },
   "author": "ysy945",
