+++ js-plugins/vue/src/farm-vue-plugin.ts
@@ -1,18 +1,17 @@
-import fs from "fs";
-import path from "path";
-import { parse } from "@vue/compiler-sfc";
-import { JsPlugin } from "@farmfe/core";
-import { handleHmr } from "./farm-vue-hmr";
-import { StylesCodeCache, CacheDescriptor } from "./farm-vue-types";
-import { genMainCode } from "./generatorCode";
-import { LessStatic } from "./less";
-import { error } from "./utils";
+import fs from 'fs';
+import path from 'path';
+import { parse } from '@vue/compiler-sfc';
+import { JsPlugin } from '@farmfe/core';
+import { handleHmr } from './farm-vue-hmr';
+import { StylesCodeCache, CacheDescriptor, LessStatic } from './farm-vue-types';
+import { genMainCode } from './generatorCode';
+import { error } from './utils';
 
 //apply style langs
-type ApplyStyleLangs = ["less"];
+type ApplyStyleLangs = ['less'];
 
 const stylesCodeCache: StylesCodeCache = {};
-const applyStyleLangs = ["less"];
+const applyStyleLangs = ['less'];
 const VueRegExp = /.vue$/;
 const JsOrTsExp = /.(ts|js)$/;
 const cacheDescriptor: CacheDescriptor = {};
@@ -21,10 +20,10 @@ export default function farmVuePlugin(options: object = {}): JsPlugin {
   //options hooks to get farmConfig
   let farmConfig = null;
   return {
-    name: "farm-vue-plugin",
+    name: 'farm-vue-plugin',
     load: {
       filters: {
-        resolvedPaths: [".vue$"],
+        resolvedPaths: ['.vue$'],
       },
       async executor(params, ctx) {
         const query: Record<string, string> = {};
@@ -36,7 +35,7 @@ export default function farmVuePlugin(options: object = {}): JsPlugin {
         const extname = path.extname(resolvedPath);
         //handle .vue file
         if (VueRegExp.test(extname)) {
-          if (vue === "true" && hash) {
+          if (vue === 'true' && hash) {
             let styleCode = stylesCodeCache[hash];
             //if lang is not "css",use preProcessor to handle
             if (applyStyleLangs.includes(lang)) {
@@ -44,13 +43,13 @@ export default function farmVuePlugin(options: object = {}): JsPlugin {
               styleCode = css;
             }
             return {
-              content: typeof styleCode === "string" ? styleCode : "",
-              moduleType: "css",
+              content: typeof styleCode === 'string' ? styleCode : '',
+              moduleType: 'css',
             };
           }
-          let source = "";
+          let source = '';
           try {
-            source = await fs.promises.readFile(resolvedPath, "utf-8");
+            source = await fs.promises.readFile(resolvedPath, 'utf-8');
           } catch (err) {
             error({
               id: resolvedPath,
@@ -92,8 +91,8 @@ export default function farmVuePlugin(options: object = {}): JsPlugin {
           );
           return {
             content:
-              "console.log(`[farm-vue-plugin]:error:there is no path can be match,please check!`)",
-            moduleType: "js",
+              'console.log(`[farm-vue-plugin]:error:there is no path can be match,please check!`)',
+            moduleType: 'js',
           };
         }
       },
@@ -101,7 +100,7 @@ export default function farmVuePlugin(options: object = {}): JsPlugin {
     // add hmr code In root file
     transform: {
       filters: {
-        resolvedPaths: [".html$"],
+        resolvedPaths: ['.html$'],
       },
       executor(params, ctx) {
         return {
@@ -114,10 +113,10 @@ export default function farmVuePlugin(options: object = {}): JsPlugin {
 }
 
 async function preProcession(styleCode: string, moduleType: string) {
-  const __default = { css: styleCode, map: "" };
+  const __default = { css: styleCode, map: '' };
   try {
     switch (moduleType) {
-      case "less":
+      case 'less':
         let lessProcessor = (await import(moduleType)) || {};
         if (lessProcessor.default) {
           lessProcessor = lessProcessor.default;
@@ -127,7 +126,7 @@ async function preProcession(styleCode: string, moduleType: string) {
         return __default;
     }
   } catch (err) {
-    error({ id: "less", message: err });
+    error({ id: 'less', message: err });
   }
   return __default;
 }
